refactor: presentation 계층 패키지 분리
